erasing type of transform functions
