+++ core/generated-sources/src/main/kotlin/org/jetbrains/kotlinx/dataframe/impl/columns/constructors.kt
@@ -150,7 +150,7 @@ internal fun <C> createColumnSet(resolver: (ColumnResolutionContext) -> List<Col
 
         override fun resolveAfterTransform(
             context: ColumnResolutionContext,
-            transform: (List<ColumnWithPath<C>>) -> List<ColumnWithPath<C>>,
+            transform: (List<ColumnWithPath<*>>) -> List<ColumnWithPath<*>>,
         ): List<ColumnWithPath<C>> {
             throw UnsupportedOperationException("Not implemented")
         }
